requirements.txt
beautifulsoup4==4.9.2
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
html5lib==1.1
httplib2==0.18.1
idna==2.10
mechanize==0.4.5
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tqdm==4.50.0
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
